requirements.txt
dotenv
streamlit
googlegenerativeai
